TCH] fix(autowatch): add guard for missing file error
Added a try/catch clause to guard against a "Path doesn't exist" error
coming from `MemoryFileSystem.prototype.readFileSync` when a new file is
added to the file-system that matches the files glob. Unfortunately, there
is not a custom type on the error, so we have to resort to string magic to
determine if that's the reason for the error.

Closes #40